Moved Data Pipeline code to new namespace
